subtrans/requirements.txt
PyQt6>=6.4.0
pytesseract>=0.3.10
mss>=9.0.1
transformers>=4.30.0
torch>=2.0.0
langdetect>=1.0.9
pynput>=1.7.6
Pillow>=9.5.0
